move FileAnalytics out
